+++ packages/orq-rc/src/orq_ai_sdk/models/createpromptop.py
@@ -1962,6 +1962,7 @@ CreatePromptProvider = Literal[
     "xai",
     "alibaba",
     "tensorix",
+    "hcompany",
     "slack",
     "orq",
 ]