+++ packages/orq-rc/src/orq_ai_sdk/models/deploymentinvokeop.py
@@ -88,6 +88,7 @@ Provider = Literal[
     "xai",
     "alibaba",
     "tensorix",
+    "hcompany",
     "slack",
     "orq",
 ]