+++ packages/orq-rc/src/orq_ai_sdk/models/deploymentsop.py
@@ -633,6 +633,7 @@ DeploymentsProvider = Literal[
     "xai",
     "alibaba",
     "tensorix",
+    "hcompany",
     "slack",
     "orq",
 ]