+++ packages/orq-rc/src/orq_ai_sdk/models/deploymentstreamop.py
@@ -2214,6 +2214,7 @@ DeploymentStreamProvider = Literal[
     "xai",
     "alibaba",
     "tensorix",
+    "hcompany",
     "slack",
     "orq",
 ]