+++ packages/orq-rc/src/orq_ai_sdk/models/getallpromptsop.py
@@ -537,6 +537,7 @@ GetAllPromptsProvider = Literal[
     "xai",
     "alibaba",
     "tensorix",
+    "hcompany",
     "slack",
     "orq",
 ]