+++ packages/orq-rc/src/orq_ai_sdk/models/getonepromptop.py
@@ -500,6 +500,7 @@ GetOnePromptProvider = Literal[
     "xai",
     "alibaba",
     "tensorix",
+    "hcompany",
     "slack",
     "orq",
 ]