+++ packages/orq-rc/src/orq_ai_sdk/models/getpromptversionop.py
@@ -536,6 +536,7 @@ GetPromptVersionProvider = Literal[
     "xai",
     "alibaba",
     "tensorix",
+    "hcompany",
     "slack",
     "orq",
 ]