+++ packages/orq-rc/src/orq_ai_sdk/models/listpromptversionsop.py
@@ -546,6 +546,7 @@ ListPromptVersionsProvider = Literal[
     "xai",
     "alibaba",
     "tensorix",
+    "hcompany",
     "slack",
     "orq",
 ]