+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KN4PGYQ5NAKR98ZRCG9812SC"
+        "01KN6CQE8R06EECMXY6M5J2Y2M"
     )
     r"""The unique identifier of the data source"""