+++ packages/orq-rc/src/orq_ai_sdk/models/updatepromptop.py
@@ -2060,6 +2060,7 @@ UpdatePromptProvider = Literal[
     "xai",
     "alibaba",
     "tensorix",
+    "hcompany",
     "slack",
     "orq",
 ]