+++ packages/orq-rc/src/orq_ai_sdk/models/updatetoolop.py
@@ -292,7 +292,7 @@ class RequestBodyTools(BaseModel):
 
     schema_: Annotated[UpdateToolRequestBodyToolsSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01KN4PGYKDHE016KKCP9XV3DF6"
+    id: Optional[str] = "01KN6CQE4TEQHAT3SJQD35KT5K"
 
     description: Optional[str] = None
 
@@ -1358,7 +1358,7 @@ class UpdateToolResponseBodyCodeExecutionTool(BaseModel):
     code_tool: UpdateToolResponseBodyCodeTool
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KN4PGYK8ZWK5JFJ4MD58T6S5"
+        "tool_01KN6CQE4HSF0JSS4BCQQJ9TEJ"
     )
 
     display_name: Optional[str] = None
@@ -1489,7 +1489,7 @@ class UpdateToolResponseBodyTools(BaseModel):
         UpdateToolResponseBodyToolsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KN4PGYK77SH4FW6AZXJPF5BA"
+    id: Optional[str] = "01KN6CQE4GPA7DNGCKEX4KCA18"
 
     description: Optional[str] = None
 
@@ -1634,7 +1634,7 @@ class UpdateToolResponseBodyMCPTool(BaseModel):
     mcp: UpdateToolResponseBodyMcp
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KN4PGYK6HMBKCC4XMXRPH6TT"
+        "tool_01KN6CQE4F6Z3BX1SQ4H4N645C"
     )
 
     display_name: Optional[str] = None
@@ -1935,7 +1935,7 @@ class UpdateToolResponseBodyHTTPTool(BaseModel):
     http: UpdateToolResponseBodyHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KN4PGYK46BXXFYT2R9GDGSTT"
+        "tool_01KN6CQE4DRJK9WJJ1JY1D2J8P"
     )
 
     display_name: Optional[str] = None
@@ -2130,7 +2130,7 @@ class UpdateToolResponseBodyJSONSchemaTool(BaseModel):
     json_schema: UpdateToolResponseBodyJSONSchema
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KN4PGYK2ZQYNZRW7D8FCDE0M"
+        "tool_01KN6CQE4CXA43BKZVWNQ04R51"
     )
 
     display_name: Optional[str] = None
@@ -2329,7 +2329,7 @@ class UpdateToolResponseBodyFunctionTool(BaseModel):
     function: UpdateToolResponseBodyFunction
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01KN4PGYK1C7M9ZR89B7FST128"
+        "tool_01KN6CQE4B56KYB914RJANT1R5"
     )
 
     display_name: Optional[str] = None