+++ packages/orq-rc/RELEASES.md
@@ -6789,3 +6789,13 @@ Based on:
 - [python v4.7.0-rc.39] packages/orq-rc
 ### Releases
 - [PyPI v4.7.0-rc.39] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.39 - packages/orq-rc
+
+## 2026-04-02 06:07:49
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.0 (2.879.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.7.0-rc.40] packages/orq-rc
+### Releases
+- [PyPI v4.7.0-rc.40] https://pypi.org/project/orq-ai-sdk/4.7.0-rc.40 - packages/orq-rc